+++ src/index.js
@@ -2,9 +2,18 @@ import React, { forwardRef } from 'react'
 
 export * from 'react-hook-form-auto'
 
-import { Autoform as RHFAutoform, setTranslations } from 'react-hook-form-auto'
+import {
+  Autoform as RHFAutoform,
+  setLanguageByName,
+  addTranslations
+} from 'react-hook-form-auto'
 
 import overrides from './skinOverride'
+
+setLanguageByName('en')
+addTranslations({
+  requiredLabel: '(required)'
+})
 
 export let Autoform = (props, ref) =>
   <RHFAutoform