+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.1
 
 * Better select control
+* Allow helper text when there are no errors
 
 ## 1.0.0